Enable unauthenticated monitoring request test case
